backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
Pillow==10.4.0
pytesseract==0.3.10
spacy==3.7.4
langchain==0.2.0
langchain-community==0.2.0
faiss-cpu==1.8.0
sentence-transformers==3.0.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
python-jose==3.3.0
passlib==1.7.4
pydantic==2.7.0
opencv-python-headless==4.9.0.80
pandas==2.2.2